app_analise_sentimentos/fase_4_kafka/backend/requirements.txt
fastapi
uvicorn
openai==0.28
pydantic
pymongo
confluent_kafka
kafka-python
python-dotenv